Selvstyrte komponenter
I dag ligger regelene som bestemmer hvilke komponenter som skal vises ganske langt unna komponentene selv.
For eksempel har vi innhold-logikk-niva3/4 som har forskjellige regler for hvilke komponenter skal vises og når.

Tanken med denne endringen er å la komponentene selv definere når de vises, og hva de viser.

Neste steg blir å fjerne niva3/4-view-laget, og heller la komponentene selv si om de krever nivå 3 eller 4.
