+++ src/ducks/autentisering.ts
@@ -1,0 +1,29 @@
+import {DataElement, STATUS} from "./api";
+import React from "react";
+
+export enum InnloggingsNiva {
+    LEVEL_1 = 'Level1',
+    LEVEL_2 = 'Level2',
+    LEVEL_3 = 'Level3',
+    LEVEL_4 = 'Level4',
+    UKJENT = 'Ukjent',
+}
+
+export interface State extends DataElement {
+    data: Data;
+}
+
+export interface Data {
+    securityLevel: InnloggingsNiva;
+    loggedIn: boolean;
+}
+
+export const initialState: State = {
+    data: {
+        securityLevel: InnloggingsNiva.UKJENT,
+        loggedIn: false
+    },
+    status: STATUS.NOT_STARTED
+}
+
+export const AutentiseringContext = React.createContext<State>(initialState);
