+++ src/komponenter/hent-initial-data/autentiseringsInfoFetcher.tsx
@@ -2,48 +2,24 @@ import * as React from 'react';
 import Innholdslaster from '../innholdslaster/innholdslaster';
 import Feilmelding from '../feilmeldinger/feilmelding';
 import { fetchData } from '../../ducks/api-utils';
-import { DataElement, STATUS } from '../../ducks/api';
 import { contextpathDittNav, erMikrofrontend } from '../../utils/app-state-utils';
 import DataProvider from './data-provider';
 import InnholdLogikkNiva4 from '../../innhold/innhold-logikk-niva4';
 import InnholdLogikkNiva3 from '../../innhold/innhold-logikk-niva3';
 import OppfolgingBrukerregistreringProvider from './oppfolging-brukerregistrering-provider';
 
+import {Data as AutentiseringData, State as AutentiseringState, initialState, InnloggingsNiva, AutentiseringContext} from '../../ducks/autentisering';
+
 export const AUTH_API = '/api/auth';
-
-export enum InnloggingsNiva {
-    LEVEL_1 = 'Level1',
-    LEVEL_2 = 'Level2',
-    LEVEL_3 = 'Level3',
-    LEVEL_4 = 'Level4',
-    UKJENT = 'Ukjent',
-}
-
-const initialState: InnloggingsInfo = {
-    data: {
-        loggedIn: false,
-        securityLevel: InnloggingsNiva.UKJENT,
-    },
-    status: STATUS.NOT_STARTED,
-};
-
-export interface Data {
-    loggedIn: boolean;
-    securityLevel: string;
-}
-
-export interface InnloggingsInfo extends DataElement {
-    data: Data;
-}
 
 const AutentiseringsInfoFetcher = () => {
 
-    const [state, setState] = React.useState<InnloggingsInfo>(initialState);
+    const [state, setState] = React.useState<AutentiseringState>(initialState);
 
     const contextpath = erMikrofrontend() ? contextpathDittNav : '';
 
     React.useEffect(() => {
-        fetchData<InnloggingsInfo, Data>(state, setState, `${contextpath}${AUTH_API}`);
+        fetchData<AutentiseringState, AutentiseringData>(state, setState, `${contextpath}${AUTH_API}`);
         // eslint-disable-next-line react-hooks/exhaustive-deps
     }, []);
 
@@ -53,14 +29,16 @@ const AutentiseringsInfoFetcher = () => {
             storrelse="XXL"
             avhengigheter={[state]}
         >
-            {state.data.securityLevel === InnloggingsNiva.LEVEL_3
-                ? <InnholdLogikkNiva3/>
-                : <OppfolgingBrukerregistreringProvider>
-                    <DataProvider>
-                        <InnholdLogikkNiva4/>
-                    </DataProvider>
-                </OppfolgingBrukerregistreringProvider>
-            }
+            <AutentiseringContext.Provider value={state}>
+                {state.data.securityLevel === InnloggingsNiva.LEVEL_3
+                    ? <InnholdLogikkNiva3/>
+                    : <OppfolgingBrukerregistreringProvider>
+                        <DataProvider>
+                            <InnholdLogikkNiva4/>
+                        </DataProvider>
+                    </OppfolgingBrukerregistreringProvider>
+                }
+            </AutentiseringContext.Provider>
         </Innholdslaster>
     );
 };
