+++ src/komponenter/registrert/registrert.tsx
@@ -1,20 +1,28 @@
-import React, { useContext, useState } from 'react';
-import { Element } from 'nav-frontend-typografi';
-import { AlertStripeInfo } from 'nav-frontend-alertstriper';
+import React, {useContext, useState} from 'react';
+import {Element} from 'nav-frontend-typografi';
+import {AlertStripeInfo} from 'nav-frontend-alertstriper';
 import Ekspanderbartpanel from 'nav-frontend-ekspanderbartpanel';
-import { BrukerregistreringContext } from '../../ducks/brukerregistrering';
-import { BrukerInfoContext } from '../../ducks/bruker-info';
-import { OppfolgingContext } from '../../ducks/oppfolging';
-import { klikkPaDineOpplysninger, seDineOpplysninger, loggAktivitet } from '../../metrics/metrics';
+import {BrukerregistreringContext} from '../../ducks/brukerregistrering';
+import {BrukerInfoContext} from '../../ducks/bruker-info';
+import {OppfolgingContext} from '../../ducks/oppfolging';
+import {klikkPaDineOpplysninger, loggAktivitet, seDineOpplysninger} from '../../metrics/metrics';
 import Opplysninger from '../innsyn/registreringsopplysninger';
 import './registrert.less';
-import { AmplitudeAktivitetsProps } from '../../metrics/amplitude-utils';
+import {AmplitudeAktivitetsProps} from '../../metrics/amplitude-utils';
+import {AutentiseringContext, InnloggingsNiva} from "../../ducks/autentisering";
 
 const Registrert = (props: AmplitudeAktivitetsProps) => {
     const brukerregistreringData = useContext(BrukerregistreringContext).data;
     const brukerinfoData = React.useContext(BrukerInfoContext).data;
     const oppfolgingData = React.useContext(OppfolgingContext).data;
+    const autentiseringData = React.useContext(AutentiseringContext).data;
     const [clickedInnsyn, setClickedInnsyn] = useState(false);
+
+    const kanViseKomponent = oppfolgingData.formidlingsgruppe === 'ARBS' && autentiseringData.securityLevel === InnloggingsNiva.LEVEL_4;
+    if (!kanViseKomponent) {
+        return null;
+    }
+
     const { amplitudeAktivitetsData } = props;
     if (!brukerregistreringData) {
         return (
